+++ packages/node-utils/server/controllers/common/apps/messages/createGetAppMessagesController.ts
@@ -9,8 +9,6 @@ export function createGetAppMessagesController(options: Options): Middleware {
       query: { merge, override = 'true' },
     } = ctx;
 
-    // @ts-expect-error 2345 argument of type is not assignable to parameter of type
-    // (strictNullChecks)
     ctx.body = await getMessagesUtil(ctx, language, appId, merge, options, override);
   };
 }