+++ packages/server/options/getAppMessages.ts
@@ -5,7 +5,7 @@ import { Op } from 'sequelize';
 
 import { AppMessages } from '../models/main/AppMessages.js';
 
-export function getAppMessages({
+export async function getAppMessages({
   app,
   context,
   language,
@@ -21,12 +21,7 @@ export function getAppMessages({
       .find((sub) => sub.type() === 'language');
 
     const baseLang = baseLanguage && String(baseLanguage).toLowerCase();
-
-    // XXX: Why do we have 3 types we could be returning here? Messages, AppMessages,
-    // and AppMessages again but defined slightly differently in `types`
-    //
-    // @ts-expect-error 2322 null is not assignable to type (strictNullChecks)
-    return AppMessages.findAll({
+    const appMessages = await AppMessages.findAll({
       order: [['language', 'desc']],
       where: {
         AppId: app.id,
@@ -38,10 +33,24 @@ export function getAppMessages({
         },
       },
     });
+
+    // XXX: Why do we have 3 types we could be returning here? Messages, AppMessages,
+    // and AppMessages again but defined slightly differently in `types`
+    //
+    // @ts-expect-error 2322 null is not assignable to type (strictNullChecks)
+    return appMessages.map((message) => ({
+      language: message.dataValues.language,
+      messages: message.messages,
+    }));
   }
 
-  // @ts-expect-error 2322 null is not assignable to type (strictNullChecks)
-  return AppMessages.findAll({
+  const allMessages = await AppMessages.findAll({
     where: { AppId: app.id },
   });
+
+  // @ts-expect-error 2322 null is not assignable to type (strictNullChecks)
+  return allMessages.map((message) => ({
+    language: message.dataValues.language,
+    messages: message.messages,
+  }));
 }