+++ packages/server/utils/render.ts
@@ -12,3 +12,22 @@ export function createGtagCode(id: string): string[] {
     `gtag("config",${JSON.stringify(id)})`,
   ];
 }
+
+/**
+ * Create a snippet of JavaScript code to initialize Meta Pixel.
+ *
+ * @param id The Meta Pixel ID to generate the code for.
+ * @returns The code to initialize Meta Pixel.
+ */
+export function createMetaPixelCode(id: string): string[] {
+  return [
+    `!function(f,b,e,v,n,t,s){if(f.fbq)return;n=f.fbq=function(){n.callMethod?
+n.callMethod.apply(n,arguments):n.queue.push(arguments)};if(!f._fbq)f._fbq=n;
+n.push=n;n.loaded=!0;n.version='2.0';n.queue=[];t=b.createElement(e);
+t.async=!0;t.src=v;s=b.getElementsByTagName(e)[0];
+s.parentNode.insertBefore(t,s)}(window,document,'script',
+'https://connect.facebook.net/en_US/fbevents.js')`,
+    `fbq('init', ${JSON.stringify(id)})`,
+    "fbq('track', 'PageView')",
+  ];
+}